DOC: Converted linux/mac doc to markdown, reformat
The Linux/Mac documentation was in MS Word, which isn't available on
Linux and is uncommon on Mac, so converted to markdown.

Reformatted pest++_input.docx, removing explicit empty paragraphs that
were used to flow the document, and changed format of tables and
paragraph styles to have MS Word automatically flow the document
according to the content.